refactor(sidebar): separate event policy from wire vocabulary
